Improved the panel, added rounded_panel
For the UI widget set, we're taking a similar approach as the Xt
intrinsics - there is a simple set function, and a simple get
function, and they act on all parameters pretty much the same.  We
can only set GLuint parameters at the moment, so the value argument
should probably become a pointer.

We decided that splitting out the rounded-corner panel was probably
the best way to go, so rounded_panel derives from panel, and adds
radius handling.

We also realized that we need a context object of some sort, to
keep track of the size of the GL pane, and to keep handles to each
of the child objects.  That's the ui::context object, in ui.h and
ui.cc.

Re: issue r9#70