Added resize handling and testing
We hadn't included the ability for the context itself to resize,
so we've added it here.  The context is a friend to the panel
because when we do a window resize, each of the children needs to
have its buffers recalculated, in order that it remains the same
size on the screen, and it seems like populate_buffers should
remain protected - that doesn't really make sense for a public
interface.