Remove the rounded panel for now
The rounded corners are going to be more of a hassle than initially
thought, and we don't have to have it now, so we'll remove them
for the time being.  They'll live on in another branch, and we can
merge them in once we figure out how we want to do things.