Improved font handling
We now handle the Freetype library entirely internally to the font.
There's no point to making the caller handle the library for us,
so we've got a couple of static functions to do it.  Also, we now
use the search paths that we added to the configdata object to
actually try to find the font filename.  We also added a cleanup
type for the glyph cache.