Fleshed out the label some more
The get and set routines are mostly done.

One of the things we're requiring is that, when setting a background
image, the width and height of the image be already set in the
object.  If either is 0, we'll throw an exception.

We're also allowing no width and height arguments to the
constructor.  Not sure if that's strictly legal, from a "hey, this
signature isn't like the others" standpoint, but we'll see how it
goes.