Move the string evaluator to the font
Kerning data is available in the font, and we'd like to include
that in our calculations.  However, that's data that should be
retained inside the font, so the string-size evaluator probably
belongs in the font, in order to keep those calculations opaque.

We might also consider putting the "make a bitmap array" function
into the font as well, so we can use the kerning data there too.