Comment on research results
We did some looking at what sort of generality we really do need,
and found some good references.  We might be able to ignore vertical
text entirely, which will massively simplify things.  However, we
will want to render L-to-R and R-to-L sections of a given string in
the right way, so that will once again massively complicate things.