Remove support for vertical fonts
We had originally thought that supporting every kind of font, in
any direction, would be a super cool thing to do.  It would have,
but it would have introduced an amazing amount of complexity, and
from some research, it seems that a lot of vertical languages allow
horizontal usage nowadays, probably for the same reasons that we're
running into.  Any special glyph handling for vertical text is no
longer available.