Make the string renderer work correctly
We had a bunch of little tiny errors which made the string
renderer not work quite right (and in many cases seg-fault).  So
we've got the math right, and things are oriented correctly in the
resulting bitmap.  There may be residual problems, and R-to-L may
not work correctly, but it mostly works for L-to-R.