Get the labels drawing correctly
We now draw the labels almost 100% correctly.  There were a few
texture parameters we neglected to set, and some packing params
that were incorrect, but we're showing both images and text on
the screen.  The only things left are to get the text into the
right color (the foreground color), and to make sure R-to-L text
and mixed-direction text function correctly.