Use GLM vectors for quadtree points
Since we've already got a nice library which does point vectors,
and can do math and transformations on them, let's go ahead and
use it.

We also recalled that we need to pass quadrant sizes into the
constructor.  We're not going to do any sort of adaptive moving of
the center point, so we can construct the tree outright before we
even put anything into it.  We don't expect to have many objects
in this tree at any time.

When we search, we want a single object that is at our point of
interest, or NULL if there is none.  We don't need to return an
entire list.  That'd just be silly.