Repair to allow client UI library to link correctly
On Travis, on the client test programs, we're running into a link
order problem - for some reason, even including libGL on the link
line with the client UI library, we're not finding any of the GL
functions that we need.  We'll try linking the library itself with
libGL et al, and maybe it'll start working.

Also a few lines that got added to config.h.in that we missed in
some previous commits.