Repair incorrect quadrant handling
We were getting random segfaults, and it ended up being the fact
that we were using the quadrant mask to index into the quadrant
array, rather than the actual index.  We need to be able to get
both the index and the mask independently, so we split them out
into their own inline functions.

Also a little renaming to make things more sensible.