Repair incorrect context handling
We weren't doing the right stuff regarding the context when we did
a resize or position change, so we're making sure to update
ourselves with the context at both of those times.

Additionally, we were adding ourselves to the context too early,
before we even had a valid position.  This may have contributed to
some of the random segfaults we were seeing.