Repair incorrect button border resizing
We were using the incorrect assumption that the border defines were
indices, rather than mask values, so we weren't setting our borders
correctly in the button grow/shrink.  Now we're shifting a bit
through the correct positions for the resize.

This addresses the root of issue r9#86, but there is still more to
fix here.