Factor the label image into a struct
We need to be able to keep track of the size of an image indepen-
dently of the size of the panel it's in, so we'll make a struct
for the image itself.  This is in preparation for not having the
border on top of the image itself.

Still working on r9#86 here.