Refactor vertex calculations
We took a strange approach in the original calculations, repeating
a lot of the calculations needlessly.  We now take the pixel sizes
at the start, and scale the margin/border sizes by them, instead of
recalculating them each time we use them.

Unfortunately (or fortunately, depending on how you look at it),
this doesn't change the problems we're seeing with the label, in
that the borders are partially over top of the images.  Though it
does tell us that our calculations were correct the first time.