Repair scaling of label images
We were seeing the label images, both image and rendered text,
scaled incorrectly, and the borders appearing over top of the
images.  It turned out that we were using the wrong scaling
factor - the real pixel size has not much to do with the mapping
of the image onto the [0, 1] range.  We use the image size to
calculate the "pixel sizes", and the images now have just the right
s/t values.

This should bring r9#86 to a close.