Repair incorrect label sizes in context
When a label has a background image, we were not seeing the right
size in the context when trying to mouse over it - it seemed that
only the image size was used to do the searching.  We now update
the context when we render the full set of buffers in the label,
immediately after we reset the label's size.