Always retain extra outside pixel in label
Previously, if the border width for a given side was nonzero, we
would add an extra pixel so that the text wouldn't bleed into the
border.  We want to retain that extra pixel all the time, in case
a borderless, marginless label abuts something which it could
also bleed into.