Add active element for button handling
We need to refactor how the button does active and armed.  We need
to be able to have a minimum margin size, so that the button does
not change size as it's activated and armed.  We'll add a member
for the active state of the button, so we can test against it.