Update ui context with key and motion call data structures
We need to be able to pass call data into our children's key
callback routines, so we'll make a structure that contains the
appropriate info.

We also noticed that there was no mouse-motion call data structure,
so we added one, for completeness.  Nothing that we're directly
doing uses it at the moment, but it could be useful in the future.

We also generalized all of the structs as typedefs, rather than
bare definitions; we may have been getting unnamed structs
allocated with the previous definitions.