Implement drawing of the text field cursor
This is the last bit of the drawing, before we actually try to use
the text field.  This commit implements the cursor, and adds the
appropriate calls in the draw routine to make it appear.

Getting close on r9#72.  Should be just the text input, and whatever
bugs we find.