Correct string drawing errors
We were not getting the string drawn into the box.  It appears that
we must always try to draw the string, even if it doesn't exist, or
we won't ever be able to draw it?

The cursor is sort of working, but it still needs some more work.

Re: issue r9#72