Repair cursor blinking
The cursor was getting turned on for one frame, which is totally
not what we want.  We restructured the draw routine to properly
draw the cursor, and also whenever we change anything about the
string or cursor, we turn the cursor on.  We also added checking
for a zero blink value, which should turn the cursor on all the
time.