Repair spaces at the end of a string in the font
When a string had a space at the end of it, the cursor would rest
at the beginning of the space, rather than the end.  This was due
to the special case in which we would use the width of the last
glyph instead of the advance; the space glyph has a zero width but
a proper advance.  We now check whether the glyph has any width
before using it as the last character.