Ensure string background is correct
In each of the string types (labels, buttons, and text fields), we
were having the trouble that the background of the string itself
was black, and the background outside of the string was somewhat
related to the background color we had set.  It looked terrible,
and that's not what we wanted.

We now have a uniform for the text background, and we mix the
foreground and background values.  Things look much better.

Re: issue r9#93