Handle longer strings than can be displayed in text field
When we are given a string that is longer than we can display at
one time, we have to figure out which part of the string to
display.

The initial idea was to figure out how many "chunks", or portions
the string needed to be chopped into, in order to display a "chunk"
in the pane.  This commit presents a solution, in that the
appropriate part of the string is shown in the pane, but it doesn't
act quite like we had in mind.  We can always amend this commit if
we get it working the way we want.