Repair the string chunking in the text field
Instead of doing anything complicated, we'll always chunk our
string up into pieces which are half the width of the widget.  We
don't get any weird half-glyph shifting now, and things act like
we want them to.  The cursor is still wrong, but at least the
behaviour of the string itself is correct now.