Repair cursor positioning in the text field
Long strings now have correct cursor positioning.  We can force the
cursor position to a particular value.

We may do regeneration of the string image and buffer data more
than is strictly necessary, but we have everything we need working,
and really this isn't going to get called a whole lot, in the grand
scheme of things.

This should conclude issue r9#96.