Add transform handling to the composite
Since the children of the composite have their position relative to
their parent, rather than the raw screen position, we need to be
able to hierarchically translate them into proper position.  We've
added a translate uniform to the shader program, so now we'll set
the composite up to use it.

Descendents of the composite will override the get_transform method
in order to translate children further into their own coordinates.

Re: issue r9#78