Add pixel size retrieval from the composite
When we draw any of our widgets, we have to use floating point
values for the points, and we always want those values to
correspond exactly with pixel values.  To do that, we need to be
able to get the exact size of each pixel, scaled to the [-1, 1]
ranges that GL uses.

We'll end up overriding this in the manager, since we just need to
chain up to the context for this call.

Re: issue r9#78