Override get/set size in row-column to add grid sizes
We can manage the size of the row-column widget in terms of grid
sizes in addition to raw pixel sizes.  We can get and set the row
and/or column size.  Once we start inserting things into the grid,
we will need to add a flag for the order in which we do the
insertion, e.g. row-major or column-major.

This is the second half of r9#78 in earnest.