Calculate grid size in row-column
We can have a full setting for the grid size in our row-column, but
that doesn't mean it's right.  We can also have 0 for either of the
two values, which means that we don't care how many elements are in
that dimension.  So we need to do some checking and possible
calculation to get correct values.

We handle the cases of:
* no grid size setting
* full grid setting, which fits all children
* one or the other of the grid settings being zero

We do not yet handle:
* full grid setting, which DOES NOT fit all children

There is an else-block for the unhandled case, but it is currently
empty.  So we'll need to put something in there.

Re: issue r9#78, part 2