Add header for popup menu
The pie menus from issue r9#66 have officially begun.  We now have a
minimal header file for the menus.

This is probably going to be a rather long series of commits.
There may need to be some refactoring in the composite widget, as
we may need to set some callbacks in the top-level context.