Repair questionable callback call data in composite
For some reason, we were passing the callback list which was
initiating the callback as the state, rather than the actual state,
in both the mouse button and key callbacks.  The manager used those
values in its callback routines, so we fixed those.

There were also a few stylistic differences we wanted to change in
the manager, while we were nearby.