Add some members and access functions to popup menu
We need to know a couple things about being a popup menu, notably
which button we should respond to, and whether we are actually
visible or not.  All of our widgets so far are just visible when
they are created, but for the popup menu we need to be created in
order to handle callbacks, but not visible when we're not needed.

We may eventually just have a visible flag everywhere, analogous
to the 'managed' state of Xt widgets.

Re: issue r9#66