Rework the popup logic in UI context
This rearranges the checking we do when we're deciding whether or
not to pop up the popup in the UI context.  We'll put off the query
into the popup itself until we decide that it will be done.  The
only expensive check is the tree search, and it's got to be done,
but we'll do it almost last.