Prepare for generating points in popup menu
It's going to be kind of complicated to generate all the points in
the pie menu, but then it was kind of complicated in the panel too,
This commit adds most of the setup for the generation.  We'll make
a minimum at 15 points around the ellipse, or a third of the width
in pixels.  This may be a bad set of defaults, but we can experi-
ment with them once things start appearing on the screen.

Re: issue r9#66