Add meaningful side definitions for popup menu
We use the regular border and margin members in the popup, but
"top", "left", etc. don't really have meaning in that context.
We use the top element for the outer border/margin and the bottom
element for the inner, so we'll add those constants to the side
namespace.

Re: issue r9#66