Update popup child creation in UI test
Since we didn't do any child handling at all in the popup menu, we
commented the child creation out, so as not to cause confusion.
Now that we do things with the children, we can uncomment the loop.
We also switched things to use labels, just to see what they looked
like.

Re: issue r9#66