Repair close codepaths in client UI
In issue r9#125, we were running into some serious segfaults, and the
few things we'd previously tried did not help.  We now add a flag
to the panel, and have the composite check its children's flags on
every button or key event.  It's not ideal, but at least we can now
close windows without the program blowing up, and also without
causing an infinite loop of some sort.

We explicitly do not perform this same check during mouse motion
events.  It's likely that at some point, somebody will actually
want that, so we'll have to do some refactoring.

One other idea that came up during this fix was to keep a to-close
list in the UI context, and move things there once they were ready
to be destroyed.  They would be out of the trees of their various
parents, and would not be drawn.  When to handle them was the big
problem:  would we have a reaper thread, or would we trigger the
cleanup on every trip through the event loop, or some other method?