Repair string renderer again
Issue r9#131 uncovered that the 'left' element of each glyph was not
being taken into account fully as we drew the string, so the
results of get_string_size did not match with the actual rendered
string.  We placed things correctly, but did not alter the
positioning of further glyphs with the appropriate placement.