Add hard string splitter to multi-line label
One of the cases we need to handle in the multi-line label is the
case in which our string segment doesn't have any whitespace we can
use to split on, so we have to make a hard split in the middle of
a word.  We'll binsearch our string for the longest thing we can
fit in our width, and do the split there.

Re: issue r9#80