Repair hard splitter in multi-line label
Our hard-splitter was not working correctly.  It turned out that
our logic for moving the endpoints was simply backward.

Additionally, we added a flag to the main splitter function to not
skip over characters if there was no whitespace.  This is not the
entire fix, but it gets us closer to where we need to be.