Flesh out the rest of the UI widget class
This commit adds the OpenGL parts, which we'll have as a part of
this widget, and will be available to all other drawn widgets.

One change that we're making here is to generate all the points as
if we were at (0, 0), and transforming them into place with a
matrix.  This will cascade down the widget tree, in that any parent
will pass in their position matrix, and children will translate
that matrix into its appropriate place.  Should cause much less
buffer regeneration.