Inherit UI composite from UI rect
We've recently learned that virtual inheritance can solve the
problem we're seeing with our multiple-inheritance points, in that
we'll only have a single copy of the rect, rather than two.  That
was mostly the goal here in the first place.

Most of the rest of the toolkit is not converted, so that will come
in subsequent commits.

Re: issue r9#123