Add border to UI widget
It is unnecessary to have the widget and panel as separate classes,
since neither will be used separately.  We'll just move things out
of the widget and into the panel, and eventually just delete the
panel.  We've got a lot of refactoring to do, so we'll do it as we
move each part.

Re: issue r9#123