Factor out search tree generation in UI composite
We need to be able to regenerate the search tree outside of any
given set call as one part of our method to properly propagate a
resize event.  As we add objects into our search tree, we first see
if they are visible, and add them only if true.

There were a couple other instances of not checking the visibility
of an object before adding it to the search tree, so those are
handled in this commit as well.

Re: issue r9#123