Remove UI panel
We no longer use need it, as all its relevant functionality is now
in the widget.  We no longer try to build it either.

Also some small rearrangement of our built modules, and addition
of a couple of files which were not previously referenced in the
Makefile.

Re: issue r9#123