Refactor widget-size calculation in UI label and text field
We were running into some cases in which the widget size was not
being properly calculated, so we now just keep all the calculation
in one place, and don't take any arguments.  Each widget knows how
big it should be, so we'll just keep that inside the method.

We were also running into cases in which the widget seemed to be
sized correctly, but had no contents.  That turned out to be an
order-of-operations thing.  To resolve that, we now override the
border- and margin-setters in the label, and add our size
calculation calls.

Re: issue r9#123