Override UI widget transform recalculation in manager
Typically the recalculation of the transform matrix is done either
when we resize, or when our parent does (i.e. the UI context).
We'll have the manager do the same for all its children at that
time.  Absent a callback list or some other mechanism to propagagte
a resize event, this is the least intrusive way I can find to do
what we want to do here.  And what's more, it actually works, which
is the root reason why we're doing this major refactor.

Re: issue r9#123