Reimplement UI widget close method
For issue r9#126, we wanted to change the unweildy close method to
instead move things up into the context where they would be
garbage-collected.  We had removed the close methods entirely, so
we have added them back, and with a pretty clean garbage-collection
scheme.

Also uncommented the comm dialog's temporary changes.  Seems to be
working correctly.