Hook the UI timeouts up in the widget draw method
We can't use a thread to handle the timeout timing, becuase all of
the GL-related stuff must be in the same thread.  We can't use a
signal to handle timeouts because who knows which thread will get
the signal.  So we'll hook our timeouts up in the draw method,
since it gets called bunches of times per second.  We won't get
hard realtime performance, but if we need that in a UI, we're
probably doing something desperately wrong.

Re: issue r9#64