Pass the widget's close argument into its parent
Now that the composite widgets all understand the sync/async flags,
we'll pass the flag we get in the call to our close method into our
parent.

We're running into some trouble with infinite looping, and
subsequent stack-blowing, so we need to do some analysis on just
how the process works.  It may be the case that we just set a
flag during resize, and turn off some of the behaviour during the
procedure.