Repair multi-line string copy-back routine
We were running into incorrect strings, and also in some cases,
segmentation faults, due to incorrect math in the portion of the
multi-line string method which copies the individual lines back
into the main image.  We were doing the right calculation to
figure out the number of lines to skip between strings, but we
were doing it at the wrong time, and as a result we were doing the
calculation with the wrong data.

There has to be a better way to do this, but it's going to be a
fairly large refactor, of a very confusing module.  Not today.
The results are visually correct, even if they aren't efficient.

Re: issue r9#137