Eliminate confusing checking in UI widget margin and border setters
We've run into bugs on a few occasions, in which the order of items
we pass into the set_va call made some or all of our borders and/or
margins not take effect, due to the size range checking in the
setters.  It makes things unnecessarily stringent, and obviously
can be the cause of subtle strange bugs, so we're just going to
take it out.