Remove redundant overridden function from UI multi-label
On reviewing the pull request for the client log entry widget,
we noticed that the calculate_widget_size method in the label and
the same function in its immediate child, the multi-label, were
identical.  No need to override something with exactly the same
thing, so we'll get rid of the override entirely.