Rework the complicated is_l_to_r() method for the font glyphs
This method has a cyclomatic complexity of 61, in addition to
being kind of terrible.  What we can do is make a lookup table of
the ranges which we understand to be r-to-l, and if our point falls
into one of those ranges, we'll know what direction it's supposed
to be.

Unfortunately the std::pair has some template functions that we
can't figure out how to override (and don't work for our needs),
so we're stuck making our own structure.  We'll call it a 'range'.

We'll put all these ranges into a std::set in a future commit.