Add autoconf configure script and support files
Instead of doing a half-dozen or more commits to get the main part
of the autoconf files in place, we're just going to do them in one
whack.

We've got an m4 module to check for perl modules which we don't
use quite yet, but we'll be doing some perly things in the test
suite, and will need some specific non-core perl modules.