Add profile-enable M4 macro
We need to be able to enable profiling when the configure switch is
passed in.  Different compilers can masquerade as gcc, but the ways
they do profiling are very different, so the switches and libraries
depend on the actual compiler, rather than simply what the CXX
environment variable is set to.