Ensure all extra libraries are being used for the LIBS variable
In some places we set LIBS directly, and in some places we set our
libraries in LDLIBS, so we should make sure they're all set into
the LIBS variable.

Also add some "this is what we are using" lines to the configure
output.  Not necessary, but helpful for debugging, or just being
aware of what's going on.