Clean up test Makefile
We were building some of the files which were already contained
within the library, so instead we'll just link the library.  That's
the code we really want to test anyway.

Also having the lib as a source of the test programs was kind of
dumb, so we got rid of it.