Add test for active which_callback
We can't really test the method directly, because it gives back a
reference.  We tried taking the address of the reference and
comparing it to the address of the member, and we were getting
different values.  We'll just add things to a callback list, and
make sure that the appropriate callback list gets picked.