Clean up composite event handlers
We want to get the main event handlers out of the callback lists,
so that we can have the callback lists independent of the event
propagation.  First, we need to make sure that a composite object
will call its own callback lists as necessary.

In future commits, we'll remove the shims from the manager, and
have the composite check whether a child is a composite before
calling either a callback list or an event handler method.