Remove callback shims from manager
We're moving the event handlers out of the callback lists, so the
shims we're using in the manager to insert the event handlers into
the callback lists need to go.

Only thing left is to get the composite's various routines to
correctly sense whether the child in question is a composite or
not, and call the appropriate method in the child.