Add more event handler functions so we can pass call data structures
Since we may be calling an event handler from within another event
handler method, and we'll already have a call data structure, there
is no need to always create a new one.  We'll have the main methods
accept the call data, and the methods with the original signatures
will do the construction of those structures.