Have CFLAGS/CXXFLAGS default to nothing if not set by the user
Autoconf defaults to '-g -O2' for its CFLAGS and CXXFLAGS variables
which is not really what we always want.  In a debugging context,
that doesn't make a whole lot of sense, and profiling can get a
little squidgy with optimization turned on.

We'll default to nothing in those variables.  We may want to add in
a --enable-release flag or something to turn on optimization and
whatever else, but we'll handle that later.