Add simple connector function for GLFW programs
We'll keep things simple for users of GLFW.  Include a single file,
compile a provided .cc file, and link it in, and it'll just work.
Our function doesn't really do any of the connecting yet, but this
commit adds the simple include file and the front-facing function
that people should call to connect our toolkit to a GLFW window.