Use new GLFW connector in our uitest program
We've got to have some sort of OS-facing toolkit, and when we came
from R9, we were using GLFW, so it's not a far stretch to continue
using it.

Eventually we'll have other connectors for other sorts of windowing
libraries, and we'll be able to specify which one to use in the
configure step, but for now this is the only one we've got.