Split the mouse button and action maps in the GLFW connector
We found that there were some GLFW values which were actually the
same, and caused the buttons to stop working correctly - the 'up'
action mapped to something that was the same as something else.

We'll split the mouse button map and the mouse action map into two
separate maps, and all will be well.