Check for Xvfb in configure
We want to run an Xvfb during our test suite, so configure should
check for it.  We'll try a bunch of the typical paths we might see
and set the variable to the full path of the program.  We'll also
have the WANT_TESTS makefile conditional depend on whether we found
an Xvfb program, along with the GLFW requirement.