Add very simple xvfb target in test makefile
In order to actually run whatever test we have in mind, we need to
start up the Xvfb with some reasonable options.  We'll have it
terminate once the last program exits, so it won't be left hanging
around after our test.  We'll also make sure to load up the XTest
extension.