Clean up perl module checking M4 file
Our perl module checker previously didn't take multiple modules
into account, so each check was performed in isolation, and the
final result was simply the result of the last check.

Instead, we'll keep counts of things, so then we can check for a
nonzero count of the various conditions, and perform the
appropriate clause.

Also, a bit of cleanup in the presentation of our checks, to mirror
the AC_CHECK_HEADERS output.