Repair the confusion in the test makefile
We were building some of our programs at questionable times, and
running other targets when we shouldn't have.  We now have two
separate variables to keep our test list and our list of programs
which we need to compile.  The important thing here is that those
two lists are NOT THE SAME.