Make the xvfb target a dependency of the uitest program
When we try to run our xvfb target as part of the test suite, it
is always going to fail, since it generates output that the test
harness doesn't understand.  The _DEPENDENCIES mechanism allows us
to run specific targets in particular circumstances, so we'll start
up the xvfb before we even compile the uitest program.  That will
definitely allow xvfb more time to start up as well.