A few Travis tweaks to help the build
We're seeing problems with loading the GLFW library from where it
gets installed, so we'll set an LD_LIBRARY_PATH to hopefully find
it.

Also, we don't need ALL of boost, which does take a significant
amount of time to install, so we'll only install the small piece
that we actually need.