Add a function to convert GL versions to GLSL versions
We're going to key our shader files for the version of GLSL that
they support, so we can load the right one for what we want to do.
We can easily get the GL version, and there's a well-known
conversion from that to GLSL version strings, so we'll add a map
containing the conversions and a function to spit out the right
string.