Construct the shader path from a type string
We can get the GLSL version, so we want to incorporate that into
the filename we try to load when we load up a shader program.
Instead of passing in the filename, we'll just pass in the shader
type that we want to load, and we'll concoct a string from it.