Only bind the fragment shader output on GL 3.0+
Research has uncovered the fact that in OpenGL 3.0, the way that
color data emerges from the fragment shader changed.  We had been
doing things in the new-style way, but since we may now be called
upon to do things in the older context, we should handle it
seamlessly.

There will almost assuredly be some refactoring that occurs with
this, as we can generalize the need for different shaders by larger
groups, instead of individual GL/GLSL versions.

[ci skip]