Refactor the shader loading again
Based on the new information we've got, we can limit the number of
shaders we need to provide to 3 each, and can simplify the logic
behind the "which version do we use" to eliminate the version map.

We'll also ensure that our values have something meaningful in
them, in case the version retrieval doesn't do anything.