Cleanups in OpenGL 3.3+ shaders
There were a few incorrect items in the GL 3.3+ shaders, notably
in (over)use of the layout specifiers.  Layout specifiers are not
needed (or legal) in the parameters between the vertex and fragment
shaders, so we dropped them entirely.

The input locations are per-object, rather than per-eleent, which
means that a vector of any size takes up only one layout position,
rather than the number of positions that corresponds with the size
of the vector.

The layout specifiers on the output from the fragment shader take
both a location and an index parameter.  The spec wasn't entirely
clear about what those numbers legitimately *mean*, but testing of
this version of the shader programs works fine, so we won't worry
that we're doing something wrong.