Interact correctly with the uitest program
There were a few things going on that we were doing incorrectly,
and were getting no coverage stats out of the uitest program.

Firstly, we were calling WaitWindowViewable in a scalar context,
so weren't getting the actual window ID.  Next, we weren't checking
any of the return statuses of the various functions we were
calling, so we had no idea whether they were doing anything or not.

Once we did that, we found out that PressReleaseKey wasn't working
at all, so our program was never terminating correctly, and as a
result, never writing out its profiling and coverage information.
SendKeys seems to work correctly now, so we're getting what we
need.