Add a get method to the row-column
For some reason, we never added a get() method to the row-column,
so the order was impossible to retrieve.

We'll also test the new method.  We can have our reorient callback
switch back and forth between its original layout and its newer one
with each click.  Should give coverage of most of the get
functionality.