Move pie menu child positioning into set_desired_size method
When we restructured the way the composite handles its children, we
never updated the pie menu, so it didn't reposition its children at
all.  The set_desired_size method is the correct place to do that,
so we'll add an overriding method.

It doesn't work quite correctly, because the children end up in
bizarre places, though scaled from the upper left.