Position sectors and children properly in pie menu
For some bizarre reason, we were using an integer to represent a
fraction of a radian, for our sector angles.  Switching to a float
made things actually appear at different and correct angles.

We also found that the locations of the children were unnecessarily
including the position of the pie menu.  This was likely a remnant
from before we used the transformation matrix to position children.

The sector boundaries are still not being drawn, but at least once
they are, they'll be in the right places.