Move upper-left of pie menu to the actual upper left of the buffer
In working to get the sectors recognized, we found that the buffer
actually contains a set of primitives which doesn't correspond with
the actual position of the widget - its center is at the upper-left
corner.

This moves the center to the center, so the parent can correctly
dispatch mouse position events.  Fortunately this change is
opaque to the widget's parent.