Factor child selection into its own method in pie menu
We need to add a mouse motion callback to our pie menu, and we'll
need a simple method to call to retrieve which widget is in a
given sector (or at a given position).  This adds the which_child
method, which does just that, and factors it out of the existing
mouse_btn_callback.