Override composite mouse_pos_callback in pie menu
Performing a search of the quadtree inside the pie menu doesn't
give us what we need, because the sectors are oddly shaped and do
not conform with the typical elements that the rest of the
composite-derived widgets expect to have as children.

As a result, we'll virtualize the mouse_pos_callback in the
composite so we can override it, and override it in the pie menu.
The function is largely the same, though we no longer worry about
our list of children changing during the callback.  Also, we no
longer have to worry about handling events on a part of the pie
menu which isn't covered by a child, because there are no such
parts in the pie menu.