Remove unused normal attribute from UI context
The coverage reports also showed that the normal attribute is
never requested.  It is a remnant of when we were trying to use
normals as a way to generate rounded corners on the widgets (it
didn't even close to work), and we never removed it.  It's never
referenced in any of the shaders, so there's no need for it.