Repair skew for elliptical pie menus
We were seeing some numerical oddness, in that with elliptical pie
menus, the sector we were getting back from the which_sector
classifier didn't match up with what we were drawing on the screen.

We figured out that circular pie menus (with width equal to height)
looked and acted just fine, so the ellipse was skewing our angle
and distance values.  We introduce a factor which normalizes the
x- and y-values, so that the classifier returns correct data for
both angle and distance.

Re:  issue #31