Move load_glyph method into base_font class
Loading a single glyph is pretty basic functionality, so it makes
sense to have it in the base class.  However, the base class will
not have its own font handle, so we'll need to have the function
take a handle as an argument.  This will enable the font-list to
have multiple individual fonts, and be able to load from any of
them as needed.