Add Freetype face handle to the glyph structure
Once we start having more than one font contained within a single
cache object, we'll need to be able to determine what font the
glyph comes from (for kerning, at the very least).  For now, we'll
set it, and ignore it for the rest of the life of the glyph.