Move kern method into base_font
The kern method is another piece of basic functionality which
should reside in the base class.  Fortunately, it doesn't need a
signature change, so existing calls can remain as-is.

Now that we have which font a given glyph comes from, we can make
sure the glyphs we're trying to kern actually come from the same
font before we call out to the library.  If not, we'll use the same
bailout condition as the FT_Get_Kerning failure.