Move utf8 <--> u32 string converters to font files
Having the string converters sequestered within the label is the
wrong place for them, as they are useless for anything but the
label.  We'll move them to be top-level functions in the ui
namespace, so they're available to everybody.  The font files are
a more reasonable location for them to live, so we'll put them
there.